backend/requirements.txt
# FastAPI & Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Queue & Async
celery==5.3.4
redis==5.0.1

# GitHub Integration
PyGithub==2.1.1

# Secret Detection
regex==2023.12.25

# Utils
requests==2.31.0
python-multipart==0.0.6
pydantic-extra-types>=2.5.2

# Logging & Monitoring
python-json-logger==2.0.7
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# CORS
fastapi-cors==0.0.6

# Rate Limiting
slowapi==0.1.9
